+++ openwave/xperiments/__init__.py
@@ -1,5 +1,5 @@
 """
 Xperiment modules for OpenWave Energy Wave Theory simulations.
 
-Don't include _archive in package documentation
+Don't include _archives in package documentation
 """
