+++ openwave/xperiments/waves_granule_medium/spherical_wave.py
@@ -1,18 +1,18 @@
 """
-XPERIMENT: 3D Spherical Quantum-Wave
+XPERIMENT: 3D Spherical Wave
 
 Run sample XPERIMENTS shipped with the OpenWave package or create your own
 Tweak universe_edge and other parameters to explore different scales.
 
-Demonstrates radial harmonic oscillation of all granules in the BCC lattice.
-All granules oscillate toward/away from the lattice center along their
-individual direction vectors, creating spherical wave interference patterns.
+Demonstrates wave interference from multiple sources in the BCC lattice.
+Each source generates spherical longitudinal waves that superpose at each granule,
+creating constructive and destructive interference patterns.
 
 This XPERIMENT showcases:
-- Radial wave propagation from lattice center
-- Phase-shifted oscillations creating wave interference
-- Uniform energy injection across all granules
-- No spring coupling (pure oscillation demonstration)
+- Multiple wave sources (up to 8, configurable positions)
+- Wave superposition and interference patterns
+- Phase control between sources (constructive/destructive interference)
+- No spring coupling (pure wave propagation)
 """
 
 import taichi as ti
@@ -23,7 +23,7 @@ from openwave.common import constants
 from openwave._io import render
 
 import openwave.spacetime.qmedium_granule as qmedium
-import openwave.spacetime.qwave_radial as qwave
+import openwave.spacetime.qwave_granule as qwave
 import openwave.validations.wave_diagnostics as diagnostics
 
 # Define the architecture to be used by Taichi (GPU vs CPU)
@@ -38,6 +38,24 @@ TARGET_GRANULES = 1e6  # target particle count, granularity (impacts performance
 
 # slow-motion (divides frequency for human-visible motion, time microscope)
 SLOW_MO = constants.QWAVE_FREQUENCY  # slows frequency down to 1Hz for human visibility
+
+# Number of wave sources for this xperiment
+NUM_SOURCES = 1
+
+# Wave Source positions: normalized coordinates (0-1 range, relative to universe edge)
+# Each row represents [x, y, z] coordinates for one source (Z-up coordinate system)
+# Only provide NUM_SOURCES entries (only active sources needed)
+sources_position = [
+    [0.5, 0.5, 0.5],  # Wave Source 0 - Center
+]
+
+# Phase offsets for each source (integer degrees, converted to radians internally)
+# Allows creating constructive/destructive interference patterns
+# Only provide NUM_SOURCES entries (only active sources needed)
+# Common patterns: 0° = in phase, 180° = opposite phase, 90° = quarter-cycle offset
+sources_phase_deg = [
+    0,  # Wave Source 0 (eg. 0 = in phase)
+]
 
 lattice = qmedium.BCCLattice(UNIVERSE_EDGE, TARGET_GRANULES)
 granule = qmedium.BCCGranule(lattice.unit_cell_edge)
@@ -57,8 +75,8 @@ def xperiment_specs():
         sub.text("QMedium: Granules in BCC lattice")
         sub.text("Granule Type: Point Mass")
         sub.text("Coupling: Phase Sync")
-        sub.text("QWave Source: 1 Harmonic Oscillator")
-        sub.text("QWave Propagation: Radial from Center")
+        sub.text(f"QWave Sources: {NUM_SOURCES} Harmonic Oscillators")
+        sub.text("QWave Propagation: Radial from Source")
 
 
 def data_dashboard():
@@ -98,14 +116,15 @@ def data_dashboard():
 
 def controls():
     """Render the controls UI overlay."""
-    global show_axis, block_slice, granule_type
+    global show_axis, block_slice, granule_type, show_sources
     global radius_factor, freq_boost, amp_boost, paused
 
     # Create overlay windows for controls
-    with render.gui.sub_window("CONTROLS", 0.85, 0.00, 0.15, 0.21) as sub:
+    with render.gui.sub_window("CONTROLS", 0.85, 0.00, 0.15, 0.24) as sub:
         show_axis = sub.checkbox("Axis", show_axis)
         block_slice = sub.checkbox("Block Slice", block_slice)
         granule_type = sub.checkbox("Granule Type Color", granule_type)
+        show_sources = sub.checkbox("Show Wave Sources", show_sources)
         radius_factor = sub.slider_float("Granule", radius_factor, 0.01, 2.0)
         freq_boost = sub.slider_float("f Boost", freq_boost, 0.1, 10.0)
         amp_boost = sub.slider_float("Amp Boost", amp_boost, 1.0, 5.0)
@@ -151,15 +170,15 @@ def normalize_granule():
 
 
 def render_xperiment(lattice):
-    """Render 3D BCC lattice with radial harmonic oscillation using GGUI's 3D scene.
+    """Render 3D BCC lattice with multiple wave sources using GGUI's 3D scene.
 
-    Visualizes all granules oscillating radially from the lattice center,
-    creating spherical wave interference patterns through phase-shifted oscillations.
+    Visualizes wave superposition from multiple sources, creating interference patterns
+    where waves constructively and destructively combine.
 
     Args:
         lattice: Lattice instance with positions, directions, and universe parameters
     """
-    global show_axis, block_slice, granule_type
+    global show_axis, block_slice, granule_type, show_sources
     global radius_factor, freq_boost, amp_boost, paused
     global normalized_position
 
@@ -167,6 +186,7 @@ def render_xperiment(lattice):
     show_axis = True  # Toggle to show/hide axis lines
     block_slice = True  # Block-slicing toggle
     granule_type = False  # Granule type coloring toggle
+    show_sources = False  # Show wave sources toggle
     radius_factor = 0.4  # Initialize granule size factor
     freq_boost = 0.1  # Initialize frequency boost
     amp_boost = 5.0  # Initialize amplitude boost
@@ -181,6 +201,14 @@ def render_xperiment(lattice):
     # block-slicing: hide front 1/8th of the lattice for see-through effect
     normalized_position = ti.Vector.field(3, dtype=ti.f32, shape=lattice.total_granules)
     normalize_granule()
+
+    # Convert phase from degrees to radians for physics calculations
+    # Conversion: radians = degrees × π/180
+    sources_phase_rad = [deg * ti.math.pi / 180 for deg in sources_phase_deg]
+
+    qwave.build_source_vectors(
+        sources_position, sources_phase_rad, NUM_SOURCES, lattice
+    )  # compute distance & direction vectors to all sources
 
     # Print diagnostics header if enabled
     if WAVE_DIAGNOSTICS:
@@ -200,15 +228,14 @@ def render_xperiment(lattice):
             last_time = current_time
             t += dt_real  # Use real elapsed time instead of fixed DT
 
-            # Apply radial harmonic oscillation to all granules
-            # All granules oscillate toward/away from lattice center along their direction vectors
-            # Phase is determined by radial distance, creating outward-propagating spherical waves
-            qwave.oscillate_granules_tocenter(
+            # Apply radial harmonic oscillation to all granules from multiple wave sources
+            # Each granule receives wave contributions from all active sources
+            # Waves superpose creating interference patterns (constructive/destructive)
+            qwave.oscillate_granules(
                 lattice.position_am,  # Granule positions in attometers
                 lattice.equilibrium_am,  # Rest positions for all granules
                 lattice.velocity_am,  # Granule velocity in am/s
-                lattice.center_direction,  # Direction vectors to center for all granules
-                lattice.center_distance_am,  # Radial distance from each granule to center
+                NUM_SOURCES,  # Number of active wave sources
                 t,
                 SLOW_MO,  # Slow-motion factor for visibility
                 freq_boost,  # Frequency visibility boost (will be applied over the slow-motion factor)
@@ -244,6 +271,14 @@ def render_xperiment(lattice):
                 normalized_position,
                 radius=normalized_radius * radius_factor,
                 color=config.COLOR_MEDIUM[1],
+            )
+
+        # Render the wave sources
+        if show_sources:
+            render.scene.particles(
+                centers=qwave.sources_pos_field,
+                radius=normalized_radius * 2,
+                color=config.COLOR_SOURCE[1],
             )
 
         # Render the scene to canvas
