+++ ship_log/week_2025-10-13.md
@@ -25,7 +25,7 @@ New 3D wave propagation demo.
 
 - **Wave Source Scaling**: Increased maximum wave sources from default to 10 sources (Oct 17-18)
 - **Variable Wave Source Support**: Refactored wave source handling to support variable count dynamically (Oct 18)
-- **Crossing Waves Experiment**: Added new `xwaves` (crossing waves) experiment with dedicated `qwave_xwaves` kernel (Oct 16)
+- **Crossing Waves Experiment**: Added new `xwaves` (crossing waves) experiment with dedicated `qwave_granule` kernel (Oct 16)
 - **Wave Interference Visualization**: Implemented multi-source interference patterns showing constructive/destructive interference
 - **UI Controls**: Added toggle for showing/hiding wave sources in visualization (Oct 17)
